refactor(core): consolidate localization APIs under localization domain
- Move all localization-related APIs into the localization folder
- Remove obsolete localization APIs
- Adopt domain-based development structure
