+++ npm/react/libs/core/src/hooks/index.ts
@@ -1,2 +1,1 @@
-export { useLocalizer, useTranslation } from './use-localizer';
 export { useEnvironment } from './use-environment';
