+++ npm/react/libs/core/src/index.ts
@@ -3,8 +3,8 @@ export * from './config/index';
 export * from './hooks/index';
 export * from './http/index';
 export * from './initializer/index';
+export * from './localization/index';
 export * from './models/index';
-export * from './providers/index';
 export * from './storage/index';
 export * from './store/index';
 export * from './utils/index';
