+++ npm/react/libs/core/src/models/index.ts
@@ -1,5 +1,4 @@
 export { type ApplicationInitializer, InitializerError } from './application';
 export type { Endpoint, Environment } from './environment';
-export type { Translations, CultureInfo } from './localization';
 export type { Awaitable, Provider } from './common';
 export type { OAuth } from './oauth';
