--- npm/react/libs/core/src/providers/index.ts
@@ -1,4 +1,0 @@
-export {
-  remoteLocalizationProvider,
-  clientLocalizationProvider,
-} from './localization-providers';
