+++ npm/react/libs/core/src/store/index.ts
@@ -1,3 +1,2 @@
 export { createStore, createStoreHook } from './axiom-store';
 export { environmentStore } from './environment-store';
-export { localizerStore } from './localizer-store';
